Change state allocation.
Instead of allocating the states on the heap, they are now allocated
as member fields, and initialized in the constructor of the state
machine.

As a bonus, all QByteArrays are now created fromRawData, removing the
need of a string copy in the constructor of QByteArray.

reducing test to few states

Change-Id: I5fe8852577304a845c509141fdde70d975f72956
Reviewed-by: Fawzi Mohamed <fawzi.mohamed@theqtcompany.com>