ID mangling for C++ names.
The XML IDs can have a couple of characters that C++ can't grok as
identifier names, so those IDs have to be mangled. We can't simply
replace all illegal characters with e.g. an underscore, because that
could easily create conflicts. For example: id="a:b" and id="a_b" would
suddenly have the same name.

So, the following patterns are used:
    "_" -> "__"
    ":" -> "_colon_"
    "-" -> "_dash_"
    "@" -> "_at_"

This mangling does not affect the objectName.

Change-Id: I2ba5e4b736fb0e7e9cc3767036f8aaba8a1876b3
Reviewed-by: Erik Verbruggen <erik.verbruggen@theqtcompany.com>