Deduplicate QScxmlStateMachine ctor and remove didChange
The didChange parameter to reachedStableState() is rather confusing as
a "true" does not denote that the new state is different from the
previous one. It just tells us that at least one microstep was
performed. This information is not very useful to users of the API, so
let's drop it.

Change-Id: Ia59e9ad98d813b151429fa0067ab26d007eb4c26
Reviewed-by: Erik Verbruggen <erik.verbruggen@theqtcompany.com>