Don't bind NULL events to _event variable.
This fixes test329.

According to scxml documentation:

"NULL has no name and
is used only in these definitions.
It never occurs in the event queues of an SCXML Processor".

"In the case of <transition> elements that do not contain
an 'event' attribute [...], the _event variable will not have a easily
predictable value since the transition is not being driven by an event.
In these cases, _event will be bound to the last event that was
matched against a transition."

Analysis of test329:
We enter the s1 state, raise event "foo" (which binds the
_event variable into "foo" event) and transition into s2 state.
When entering s2 we assign the _event variable (which is
still bound to "foo") to the Var2. Next, we select
eventless transitions for s2. In current implementation,
selecting eventless transitions bind _event variable
to the empty event. Now we omit it, according to docs.

Change-Id: Ib393d4465f9a3ff4afc152aee2c334a80ee5fe66
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>