Restore Q_PROPERTY macros
This is still almost for free, and vastly easier to do than a solution
with plugins.qmltypes. The plugins.qmltypes format is tailored for
actual plugins, identifyable by a fixed URL. However, in the workflow we
suggest right now, the URL is controlled by the user. Resolving it from
the code would require the same guesswork as we have to do to resolve
Q_PROPERTY macros from the headers.

Change-Id: Iec916e7e91de096a38b61e6b76794b3dca3139d5
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>