Examples: Use QML states for trafficlights
This allows checking the traffic ui states in the designer
and makes the redGoingGreen explicit also in the ui.

While this is more code for the states I think this is
nice pattern to improve maintainabilty.

The button.source property shows that QML states are not required and
that properties can directly use the state machine.

Change-Id: I38d74f8320e27bf1bc5e8125395d907fafa44723
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>