Provide a simplified traffic light example
The traffic light QML examples are very large for the simple task they
perform. Of course we want nicely designed examples that follow all the
patterns we recommend for writing QML. However, we also want to show
the simplest way to bind states to QML properties. This is why we need
a simplified example in addition to the others.

Change-Id: I37022db162353c9bac4bfd775571d4afb666b357
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>