Simplify QScxmlParser
There is little point in exposing the intermediate steps of parsing,
instantiating the state machine, and instantiating the data model to
the user. One method to do all of this is enough and relieves us from
the burden of maintaining binary compatibility for multiple poorly
named extra methods.

Also, drop addError(). Who wants to add errors to the parser from the
outside?

Change-Id: I4a2faba8c5f130e8a95675ebbdb747df7afdee73
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>