Update plugins.qmltypes
Also, have the extended state machine type follow the usual naming
scheme, and increment the minor version everywhere. Since we don't
need to be compatible with 5.7 we don't need to drag the 5.7 version
around at all. In addition, we also don't need to namespace QtScxml
in any QML context and, using the additional type information, we can
easily constrain the types of all state machiens in QML.

The fromFile() and fromData() methods don't have to be Q_INVOKABLE and
that just confuses qmlplugindump, do drop Q_INVOKABLE there.

Change-Id: I7bd3f10ac458d64fcb98caece770b5d4c61abdb5
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>