Distinguish invalid state/transition types from IDs and drop root state
Types and IDs are conceptually different and shouldn't be mixed up this
way. Also, as far as I can see, there is no real need to invent a
synthetic root state. Leaving that out saves us a lot of inconsistency
(e.g. the root state didn't show up in allStates(), and you couldn't
ask it for stateChildren()). Without the root state some transitions
originate from the invalid state, but that is an equally good
characterization for initial transitions as the synthetic root
state.

Change-Id: I3738df08d0e3702433dabd3fb837f0ddaea57465
Reviewed-by: Jan Arne Petersen <jan.petersen@kdab.com>
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>