Drop indirect access to tableData from various data models
As we need to access it from everywhere, we can as well make it a
public property of the state machine. This saves us the tableData()
method on QScxmlDataModel and some back and forth between d and q
pointers. We can also drop the indirection for accessing the state
machine in most cases, as that is just a public member of
QScxmlDataModelPrivate.

Change-Id: I4aa0cec7d4664aec7b1581de531cdb1554ca5342
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>