Fix emission of state changed signals
The state index is not the same as its changed signal index.
The changed signal is not being generated for history states.
The current fix is that we create a hash cache of state index
into the signal index of its appropriate signal just
after we set the table data for the state machine.

Task-number: QTBUG-61243
Change-Id: I778adaf6c2d626be17c50b558f93ec0035ea3325
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>