Only return uniquely named services from invokedServices()
Previously, we used the fact that QVariantMap is in fact a multimap in
order to squeeze services with equal names into it. This doesn't work
anymore.

[ChangeLog][Important Behavior Changes] If you have multiple equally
named children of an invoked service, only one of them will be included
in the "children" property now. Before, the fact that QVariantMap was
actually a multimap was used to squeeze all of them in.

Change-Id: I12aa5dda18629a8319c425fad1b9e82e0083befe
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>