Fix flaky autotest with a delayed event in scion
Before the historical fix for the issue
(commit 3da30ee8928ab27249a74966ed4981766d62ff25)
the playEvent() method submitted directly the
delayed event and it tried to verify the configuration
just after. It couldn't work, since
the event was delayed, so the demanded configuration
couldn't be reached yet - it was checked too early.

After the mentioned fix, we didn't submit an event
immediately, but waited demanded time (in hope,
that statemachine reaches the configuration expected
after a delay) to mimic the delay, and after that,
we submitted a non-delayed event directly.
This could also cause issues like
these mentioned in task description. The reason could
be a race between the point in time when statemachine
issues the delayed "s" event and the point in time
when we submit the "t2" event. It could happen, that
we submit "t2" while still being in "b" state (so before
the statemachine issues the "s" event), and in this
case the "t2" event doesn't trigger any change in
the statemachine, thus the configuration stay unchanged.

In the current fix we go back to the original idea of
submitting the delayed event directly, but we are
delaying the point in time when we do check the configuration
expected after the event is dispatched. So is case of delayed
event we connect to the statemachine's signal emitted
when the event is being dispatched.
After the connection is established, we submit a delayed event
and start awaiting for the right moment when we should
expect the stabilized configuration. This will be signalled
with a timeout of trigger timer object.
When the submitted event has been dispatched (after the
desired delay), we connect synchronously to state machine's
reachedStableState() signal. When reachedStableState() is
received, we trigger the timer to unwind from synchronously
called slot and to get back to the place where we were
awaiting for timer to be triggered. Now we should expect
the right configuration.

The above approach works nicely also in case of playing
non-delayed events.

Fixes: QTBUG-83420
Pick-to: 5.15
Change-Id: Ia6ddbc6ac9282a2dd41e0c99dd1bb59610690c95
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>