Mark dynamic state machines' metaobjects as fully dynamic
The memory is allocated using malloc() and will be free()'d eventually.
Client code cannot expect to keep string data from those metaobjects
around.

The QML engine will avoid the creation of property caches (which hold
string data from the metaobject) if the metaobject is fully dynamic.
Therefore, just create a QAbstractDynamicMetaObject that wraps the
original metaobject and pass that to QObjectPrivate in order to signal
the fully dynamic nature of the metaobject to QML.

Fixes: QTBUG-89521
Pick-to: 5.15
Change-Id: I349dcb689be5674679237a551ed8595bd6b5b672
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>