Qt State Machine standalone documentation
The Qt State Machine was carved off of the qtbase and qtdeclarative repositories
and relocated to the qtscxml repository as a neighbor of the scxml state machine.
As such the Qt State Machine had several fragments of existing documentation, but
no qdoc packaging to make them work as a standalone documentation. This packaging
and related necessary changes is the essence of this commit.

Pick-to: 6.1
Task-number: QTBUG-89833
Change-Id: I43d8f8145d4467bd48dbf4674fa28a29671de868
Reviewed-by: Paul Wicking <paul.wicking@qt.io>