Enable non-gui builds of qtscxml and statemachine examples
This commit enables building the modules against Qt with -no-gui set.
While library compilations were possible, the examples failed to
compile as QML libraries may be present without gui. This is
different root cause than in the original bug item due to the
qmake => cmake migration, but end goal of this commit is the same.

Additionally the few examples without gui dependency are now built.
For clarity: most auto tests link against gui and are thus not 'enabled'
as part of this commit.

Pick-to: 6.1
Task-number: QTBUG-60045
Change-Id: Ifc184eb95aa465f6e7ab832b4c13017f81a47e88
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>