Enable StateMachine QML tests as autotests and add one testcase
The QML autotests are now built and run as part of the autotests,
where they previously relied on running  with qmltestrunner
manually.

Also fixes one testcase warning about missing 'initialState'.

Additionally this commit adds one QML testcase which tests a Scxml
bug that was specific to Qt5; while the bug does not happen with
Qt6 it is important to add the test case also to Qt6 to catch any
regression in the future.

Change-Id: I8fbaa44dab52b64a4d2c5c73ad5d3e3ff2eb0346
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>