Fix naming of plugin classes, targets and files
The names should be the same so that the static version of the plugin
can be included. It's also good to follow the naming conventions
established by other plugins. Also, avoid multiple inheritence on the
plugin interface by making the interface itself a QObject. Furthermore,
QScxmlDataModelPlugin, being a private class, should be privately
exported.

Fixes: QTBUG-93458
Task-number: QTBUG-93444
Pick-to: 6.1
Change-Id: Ifbb3b062c922b79d8e55469349a64304a7a3fdd2
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>