CMake: Fix iOS app builds that use QScxmlEcmaScriptDataModelPlugin
The QScxmlEcmaScriptDataModelPlugin project was setting the
PRIVATE_MODULE_INTERFACE option which is not a recognized option for
plugins, but because it came after LIBRARIES, the actual option name
and the values that come after it got recorded as "regular" libraries.

This in turn was passed to qt_internal_extend_target which /does/
recognize PRIVATE_MODULE_INTERFACE, but because there's no counterpart
Private target for the plugin target (as would be needed for
PRIVATE_MODULE_INTERFACE) the option and its values were silently
discarded.

Because qt_internal_add_plugin adds Qt::PlatformPluginInternal after
arg_LIBRARIES (which had PRIVATE_MODULE_INTERFACE as an option),
that means the Qt::PlatformPluginInternal dependency
was also interpreted as to be used for the Private target and thus
discarded.

This means none of the bitcode flags meant to be applied to the plugin
target were applied and linking of an application failed with

 ld '/qt/ios/plugins/scxmldatamodel/libqscxmlecmascriptdatamodel.a(mocs_compilation_Release.cpp.o)' does not contain bitcode.

Remove the invalid PRIVATE_MODULE_INTERFACE option.

Amends 9251d5c76fb57674e8b7d3491f064814c5e08b20

Pick-to: 6.1 6.2
Fixes: QTBUG-95921
Task-number: QTBUG-95208
Change-Id: Iaa752aa29295464957934f6d9161c0b31b3970b8
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>