Partially update moc copy from qtbase
This applies qtbase/b8c2a0c18a0676595946b5543ff88492a5fc7876 to the
copy of the moc code used in qscxmlc. The moc copy here is severely
out of sync with the qtbase version, so just running update_moc.sh
produces a lot of conflicts, which is why I only applied this one
patch.

Created QTBUG-109556 to keep track of the rest.

Fixes: QTBUG-109235
Pick-to: 6.5 6.4 6.2 5.15
Change-Id: I9e033453928110ac2aab1ecd6cde824c6abe0a97
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>