Doc: Remove warninglimit .qdocconf variable
The most common limit for the maximum number of allowed documentation
warnings is zero. As all Qt module docs include() a common configuration
from qtbase, it's better to control the base warninglimit from a central
location in qtbase/doc/global.

This allows for a temporary increase of the limit across all modules as
needed - for example, when updating the QDoc binary that the CI
provisions to a version that introduces new types of documentation
warnings.

Task-number: QTBUG-113326
Change-Id: I7c60fd10804e7309f38020e97e7f6eb32d7b4b2c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>