Update the moc copy
We need to explicitly cast the string data to char now because we want
to use the moc helper functions that expect char*. Previously we got
away with unsigned char* by declaring the members ourselves. Since we
always want to cast to char anyway when retrieving the data, it doesn't
matter.

Task-number: QTBUG-116255
Change-Id: I1ca017b5b434942606675514eddceb1f725a8bd7
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>