SignalTransition: fix binding loops
This one is a bit special, because the setter assumes that there is a
QML engine handling the object.
As a result, creating a helper object for testing binding loops is a
bit tricky. Do it by having a helper QQmlComponent based on a qml file
with a SignalTransition object as a root element.
Fix the binding loop in the setter in a usual way.

Task-number: QTBUG-116542
Pick-to: 6.6 6.5
Change-Id: Ibd22dee0619a69b52901e9fe2145fcfbd9dcf98c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>