Add a note about QScxmlStateMachine restart
As part of QTBUG-109371 it was discovered that calling start() after
stop() doesn't result in a "clean" initial step which is contorary
to the existing documentation.
It was agreed that ATM changing documentation to better reflect the
current state is more suitable, than doing a proper "clean".

This commit addresses the documentation issue

Fixes: QTBUG-109371
Change-Id: Iaec8f31b39873a1cdc2809e5773fcaa1e6f5d37c
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>